% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src-snowflakedb.R
\name{atomic_copy}
\alias{atomic_copy}
\title{Atomically copy file into database table}
\usage{
atomic_copy(con, from, to, stage_name, temporary = FALSE)
}
\arguments{
\item{con}{a connection obj.}

\item{from}{a string of the filepath to copy.}

\item{to}{a string of the table name to copy into.}

\item{stage_name}{a string of the stage name (including folders) where file will be PUT before loading.
Eg, this could be a location on AWS's S3.}

\item{temporary}{a logical.
If TRUE, then if a table is created, it is created as a temp table.}
}
\description{
atomic_copy() accomplishes atomicity by creating a intermediate table,
loading into that intermediate table, then
if the final table already exists,
   swapping the intermediate with the final table (then dropping the old final table),
else
   renaming the intermediate to the final table name.
}
\details{
Note that Snowflake automatically commits pending session transactions when
a DDL query is performed.
}
