% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src-snowflakedb.R
\name{copy_to_from_staged_files}
\alias{copy_to_from_staged_files}
\title{Copies each staged file into table.}
\usage{
copy_to_from_staged_files(con, table_name, stage_with_prefix, mode,
  FILES_METADATA, temporary = FALSE, delim = NULL, df)
}
\arguments{
\item{con}{A connection obj.}

\item{table_name}{a string of the table name to be copied into.}

\item{stage_with_prefix}{a string of the stage + files' prefix.}

\item{mode}{a string of 'safe', 'overwrite', or 'append'.
If 'safe' (the default), then no table change will occur if table already exists.
If 'overwrite', then the table will be recreated.
If 'append', then the table will be appened to if it already exists.
If arg overwrite is TRUE, then mode will be set to 'overwrite'.}

\item{temporary}{a logical indicating whether or not created tables are temporary.}

\item{delim}{a string indicating the files' delim.  Guessed if not provided.}

\item{df}{a data.frame to be used to create table (if applicable).  Defaults to NULL.}
}
\description{
If table does not exist,
   then creates table, downloading a file to determine data types.
}
\details{
Depends on S3 utils.
}
