% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src-snowflakedb.R
\name{db_load_from_file}
\alias{db_load_from_file}
\title{Load file into database table}
\usage{
db_load_from_file(con, table_name, file_path, stage_dir, ...)
}
\arguments{
\item{con}{A connection obj.}

\item{table_name}{A string of the database table name being loaded.}

\item{file_path}{A string of the file_path to be loaded.}

\item{stage_dir}{A string of the stage destination, including folder.
Eg, '@warehouse.stages.foo_stage/foo/bar'.}
}
\description{
If the table does not exist, an err is thrown.
}
