% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{guess_delim}
\alias{guess_delim}
\title{guess_delim() returns the delim for filepath.}
\usage{
guess_delim(filepath)
}
\arguments{
\item{filepath}{a string of a filepath.}
}
\value{
a string of filepath's delim.
}
\description{
guess_delim() returns the delim for filepath.
}
\examples{

guess_delim('foobar.csv.gz')
--> ','
}
