% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_sql_col_idx_exprs}
\alias{make_sql_col_idx_exprs}
\title{Make sql exprs of from's col idxs mapped to to's cols.}
\usage{
make_sql_col_idx_exprs(to_colnames, from_colnames)
}
\arguments{
\item{to_colnames}{a character vector of to's colnames, ordered.}

\item{from_colnames}{a character vector of from's colnames, ordered.}
}
\description{
make_sql_col_idx_exprs() makes sql exprs of from's col idxs mapped to to's cols.
See examples.
}
\examples{
to_colnames <- c('X', 'Y')
from_colnames <- c('Y', 'X')

actual <- make_sql_col_idx_exprs(
  to_colnames,
  from_colnames
)

expect_equal(
  actual,
  c('$2', '$1')
)
}
