% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src-snowflakedb.R
\docType{data}
\name{NOT_NULLS}
\alias{NOT_NULLS}
\title{dplyr backend support for SnowflakeDB (https://snowflake.net)}
\format{An object of class \code{character} of length 8.}
\usage{
NOT_NULLS
}
\description{
Use \code{src_snowflakedb} to connect to an existing Snowflake database,
and \code{tbl} to connect to tables within that database.
}
\keyword{datasets}
