% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{sf_connect}
\alias{sf_connect}
\title{Connect to Snowflake using a profile name}
\usage{
sf_connect(name = NULL, ...)
}
\arguments{
\item{name}{Profile name from connections.toml.}

\item{...}{Additional arguments passed to dbConnect.}
}
\value{
A SnowflakeConnection.
}
\description{
Convenience wrapper around \code{dbConnect(Snowflake(), name = ...)}.
}
