% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{Snowflake}
\alias{Snowflake}
\title{Create a SnowflakeDriver instance}
\usage{
Snowflake()
}
\value{
A SnowflakeDriver singleton.
}
\description{
Create a SnowflakeDriver instance
}
\examples{
\dontrun{
drv <- Snowflake()
con <- dbConnect(drv, account = "myaccount", name = "default")
}
}
