% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R, R/transactions.R
\docType{class}
\name{SnowflakeConnection-class}
\alias{SnowflakeConnection-class}
\alias{show,SnowflakeConnection-method}
\alias{dbIsValid,SnowflakeConnection-method}
\alias{dbDisconnect,SnowflakeConnection-method}
\alias{dbGetInfo,SnowflakeConnection-method}
\alias{dbSendQuery,SnowflakeConnection,character-method}
\alias{dbSendStatement,SnowflakeConnection,character-method}
\alias{dbGetQuery,SnowflakeConnection,character-method}
\alias{dbExecute,SnowflakeConnection,character-method}
\alias{dbGetQueryArrow,SnowflakeConnection,character-method}
\alias{dbSendQueryArrow,SnowflakeConnection,character-method}
\alias{dbListTables,SnowflakeConnection-method}
\alias{dbExistsTable,SnowflakeConnection,character-method}
\alias{dbListFields,SnowflakeConnection,character-method}
\alias{dbListObjects,SnowflakeConnection-method}
\alias{dbReadTable,SnowflakeConnection,character-method}
\alias{dbWriteTable,SnowflakeConnection,character,data.frame-method}
\alias{dbCreateTable,SnowflakeConnection,character-method}
\alias{dbAppendTable,SnowflakeConnection,character-method}
\alias{dbRemoveTable,SnowflakeConnection,character-method}
\alias{dbQuoteIdentifier,SnowflakeConnection,character-method}
\alias{dbQuoteString,SnowflakeConnection,character-method}
\alias{dbQuoteString,SnowflakeConnection,SQL-method}
\alias{dbQuoteIdentifier,SnowflakeConnection,SQL-method}
\alias{dbUnquoteIdentifier,SnowflakeConnection,SQL-method}
\alias{dbQuoteLiteral,SnowflakeConnection-method}
\alias{dbDataType,SnowflakeConnection-method}
\alias{dbBegin,SnowflakeConnection-method}
\alias{dbCommit,SnowflakeConnection-method}
\alias{dbRollback,SnowflakeConnection-method}
\alias{dbWithTransaction,SnowflakeConnection-method}
\title{SnowflakeConnection}
\usage{
\S4method{show}{SnowflakeConnection}(object)

\S4method{dbIsValid}{SnowflakeConnection}(dbObj, ...)

\S4method{dbDisconnect}{SnowflakeConnection}(conn, ...)

\S4method{dbGetInfo}{SnowflakeConnection}(dbObj, ...)

\S4method{dbSendQuery}{SnowflakeConnection,character}(conn, statement, params = NULL, ...)

\S4method{dbSendStatement}{SnowflakeConnection,character}(conn, statement, params = NULL, ...)

\S4method{dbGetQuery}{SnowflakeConnection,character}(conn, statement, params = NULL, ...)

\S4method{dbExecute}{SnowflakeConnection,character}(conn, statement, params = NULL, ...)

\S4method{dbGetQueryArrow}{SnowflakeConnection,character}(conn, statement, params = NULL, ...)

\S4method{dbSendQueryArrow}{SnowflakeConnection,character}(conn, statement, params = NULL, ...)

\S4method{dbListTables}{SnowflakeConnection}(conn, ...)

\S4method{dbExistsTable}{SnowflakeConnection,character}(conn, name, ...)

\S4method{dbListFields}{SnowflakeConnection,character}(conn, name, ...)

\S4method{dbListObjects}{SnowflakeConnection}(conn, prefix = NULL, ...)

\S4method{dbReadTable}{SnowflakeConnection,character}(conn, name, ...)

\S4method{dbWriteTable}{SnowflakeConnection,character,data.frame}(
  conn,
  name,
  value,
  overwrite = FALSE,
  append = FALSE,
  row.names = FALSE,
  ...
)

\S4method{dbCreateTable}{SnowflakeConnection,character}(conn, name, fields, ..., row.names = NULL, temporary = FALSE)

\S4method{dbAppendTable}{SnowflakeConnection,character}(conn, name, value, ..., row.names = NULL)

\S4method{dbRemoveTable}{SnowflakeConnection,character}(conn, name, ...)

\S4method{dbQuoteIdentifier}{SnowflakeConnection,character}(conn, x, ...)

\S4method{dbQuoteString}{SnowflakeConnection,character}(conn, x, ...)

\S4method{dbQuoteString}{SnowflakeConnection,SQL}(conn, x, ...)

\S4method{dbQuoteIdentifier}{SnowflakeConnection,SQL}(conn, x, ...)

\S4method{dbUnquoteIdentifier}{SnowflakeConnection,SQL}(conn, x, ...)

\S4method{dbQuoteLiteral}{SnowflakeConnection}(conn, x, ...)

\S4method{dbDataType}{SnowflakeConnection}(dbObj, obj, ...)

\S4method{dbBegin}{SnowflakeConnection}(conn, ...)

\S4method{dbCommit}{SnowflakeConnection}(conn, ...)

\S4method{dbRollback}{SnowflakeConnection}(conn, ...)

\S4method{dbWithTransaction}{SnowflakeConnection}(conn, code, ...)
}
\arguments{
\item{object}{A \linkS4class{SnowflakeConnection} object (for \code{show}).}

\item{dbObj}{A \linkS4class{SnowflakeConnection} object.}

\item{...}{Additional arguments (currently ignored).}

\item{conn}{A \linkS4class{SnowflakeConnection} object.}

\item{statement}{SQL statement string.}

\item{params}{Optional parameter list for parameterized queries.}

\item{name}{Table name (character).}

\item{prefix}{An \code{Id} object indicating the hierarchy level to list
(NULL for databases, 1-component for schemas, 2-component for tables).}

\item{value}{A data.frame.}

\item{overwrite}{Logical. Drop existing table first?}

\item{append}{Logical. Append to existing table?}

\item{row.names}{Ignored (always FALSE for Snowflake).}

\item{fields}{A data.frame or named character vector of column types.}

\item{temporary}{Logical. Create a temporary table?}

\item{x}{Character to quote.}

\item{obj}{An R object to map to a Snowflake SQL type.}

\item{code}{Code to execute within the transaction.}
}
\description{
An S4 class representing a connection to Snowflake via the SQL API v2.
Created by \code{\link[=dbConnect]{dbConnect()}} with a \code{\link[=Snowflake]{Snowflake()}} driver.
}
\section{Slots}{

\describe{
\item{\code{account}}{Account identifier.}

\item{\code{user}}{Username.}

\item{\code{database}}{Default database.}

\item{\code{schema}}{Default schema.}

\item{\code{warehouse}}{Default warehouse.}

\item{\code{role}}{Default role.}

\item{\code{.auth}}{Auth list (type, token, token_type, plus params for refresh).}

\item{\code{.state}}{Environment holding mutable state (valid, in_transaction, session_info).}
}}

