% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\docType{class}
\name{SnowflakeDriver-class}
\alias{SnowflakeDriver-class}
\alias{dbGetInfo,SnowflakeDriver-method}
\alias{dbIsValid,SnowflakeDriver-method}
\alias{dbUnloadDriver,SnowflakeDriver-method}
\alias{show,SnowflakeDriver-method}
\alias{dbDataType,SnowflakeDriver-method}
\alias{dbConnect,SnowflakeDriver-method}
\alias{dbCanConnect,SnowflakeDriver-method}
\title{SnowflakeDriver}
\usage{
\S4method{dbGetInfo}{SnowflakeDriver}(dbObj, ...)

\S4method{dbIsValid}{SnowflakeDriver}(dbObj, ...)

\S4method{dbUnloadDriver}{SnowflakeDriver}(drv, ...)

\S4method{show}{SnowflakeDriver}(object)

\S4method{dbDataType}{SnowflakeDriver}(dbObj, obj, ...)

\S4method{dbConnect}{SnowflakeDriver}(
  drv,
  account = NULL,
  user = NULL,
  token = NULL,
  private_key_path = NULL,
  authenticator = NULL,
  database = "",
  schema = "",
  warehouse = "",
  role = "",
  name = NULL,
  ...
)

\S4method{dbCanConnect}{SnowflakeDriver}(drv, ...)
}
\arguments{
\item{dbObj}{A \linkS4class{SnowflakeDriver} object.}

\item{...}{Additional arguments (ignored).}

\item{drv}{A SnowflakeDriver, or missing (uses default).}

\item{object}{A \linkS4class{SnowflakeDriver} object (for \code{show}).}

\item{obj}{An R object to map to a Snowflake SQL type.}

\item{account}{Snowflake account identifier (e.g. "myaccount").}

\item{user}{Snowflake username.}

\item{token}{Explicit bearer token (PAT or session token).}

\item{private_key_path}{Path to PEM private key for JWT auth.}

\item{authenticator}{Auth method ("SNOWFLAKE_JWT", "OAUTH", etc.).}

\item{database}{Default database.}

\item{schema}{Default schema.}

\item{warehouse}{Default warehouse.}

\item{role}{Default role.}

\item{name}{Profile name from connections.toml.}
}
\description{
An S4 class representing the Snowflake DBI driver.
}
