% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{sqlData}
\alias{sqlData}
\title{Convert a data.frame to SQL-ready strings}
\usage{
sqlData(con, value, row.names = FALSE, ...)
}
\arguments{
\item{con}{A \linkS4class{SnowflakeConnection} object.}

\item{value}{A data.frame.}

\item{row.names}{Ignored.}

\item{...}{Additional arguments (ignored).}
}
\value{
A data.frame of character columns.
}
\description{
Convert a data.frame to SQL-ready strings
}
